添加几种投掷物实体，添加部分热力学游戏内容

+++ src/main/java/com/osir/terrametalcraft/api/capability/IHeatable.java
@@ -16,7 +16,7 @@ public interface IHeatable {
 		return this.getTemperature(ThermoUtil.ATMOSPHERIC_PRESSURE);
 	}
 
-	default double getTemperature(float pressure) {
+	default double getTemperature(double pressure) {
 		return this.getPhasePortrait().getTemperature(pressure, this.getEnergy());
 	}
 }
