+++ src/main/java/com/osir/terrametalcraft/api/thermo/IPhasePortrait.java
@@ -1,9 +1,9 @@
 package com.osir.terrametalcraft.api.thermo;
 
 public interface IPhasePortrait {
-	double getTemperature(float pressure, double energy);
+	double getTemperature(double pressure, double energy);
 
-	double getEnergy(float pressure, double temperature);
+	double getEnergy(double pressure, double temperature);
 
-	Phase getPhase(float pressure, double energy);
+	Phase getPhase(double pressure, double energy);
 }
